Configure Thymeleaf layout dialect, set up existing html to use this. Move SecurityConfiguration in /config
