+++ src/main/java/com/makersacademy/acebook/config/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.makersacademy.acebook;
+package com.makersacademy.acebook.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
