+++ src/main/java/com/makersacademy/acebook/config/ThymeleafConfiguration.java
@@ -1,0 +1,13 @@
+package com.makersacademy.acebook.config;
+
+import nz.net.ultraq.thymeleaf.layoutdialect.LayoutDialect;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ThymeleafConfiguration {
+    @Bean
+    public LayoutDialect layoutDialect() {
+        return new LayoutDialect();
+    }
+}
