+++ src/main/resources/templates/layout.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en"
+      xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+<head>
+    <meta charset="UTF-8">
+    <title>Acebook</title>
+    <link rel="stylesheet" href="/main.css"/>
+</head>
+<body>
+    <main>
+        <!-- add nav here, or as a fragment -->
+        <div layout:fragment="content"></div>
+    </main>
+</body>
+</html>
