+++ src/main/resources/templates/posts/index.html
@@ -1,27 +1,23 @@
 <!DOCTYPE html>
-    <html xmlns:th="http://www.thymeleaf.org"
-          xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5" >
-    <head lang="en">
-        <meta charset="UTF-8"/>
-        <title>Acebook</title>
-        <link rel="stylesheet" href="/main.css" />
-    </head>
-    <body>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorate="~{layout}">
 
-    <h1>Posts</h1>
+    <div layout:fragment="content">
+        <h1>Posts</h1>
 
-    <div sec:authorize="isAuthenticated()">
-        Signed in as <span sec:authentication="name"></span>
+        <div sec:authorize="isAuthenticated()">
+            Signed in as <span sec:authentication="name"></span>
+        </div>
+
+        <form action="#" th:action="@{/posts}" th:object="${post}" method="post">
+            <p>Content: <input type="text" th:field="*{content}" /></p>
+            <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
+        </form>
+
+        <ul th:each="post: ${posts}">
+            <li th:text="${post.content}" />
+        </ul>
     </div>
-
-    <form action="#" th:action="@{/posts}" th:object="${post}" method="post">
-        <p>Content: <input type="text" th:field="*{content}" /></p>
-        <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
-    </form>
-
-    <ul th:each="post: ${posts}">
-        <li th:text="${post.content}" />
-    </ul>
-
-    </body>
 </html>
